+++ src/app/ui/weather-card/weather-card.tsx
@@ -1,5 +1,5 @@
 import { ApiResponse } from "@/app/lib/types/api-response";
-import { capitalizeFirstLetter } from "@/app/lib/hooks/capitalizeFirstLetter";
+import { capitalizeFirstLetter } from "@/app/lib/hooks/capitalize-first-letter";
 import Image from "next/image";
 import { getDayOfWeek } from "@/app/lib/hooks/getDayOfWeek";
 import DropHalfBottom from "../icons/drop-half-bottom";
